implementing destroy for ethernet stream
git-svn-id: svn://sidvind.com/project/BTH/NPLDP/libcap_utils/trunk@1590 8459abcf-8237-4014-9f1e-057f68ddaf13

+++ caputils/stream.h
@@ -11,7 +11,7 @@ typedef int (*fill_buffer_callback)(struct stream* st);
 /**
  * Stream destructor.
  */
-typedef int (*destroy_callback)(struct stream* st);
+typedef long (*destroy_callback)(struct stream* st);
 
 typedef int (*write_callback)(struct stream* st, u_char* data, size_t size);
